Add metadata to all plugins.
This is needed for the upcoming update system. With this system, people will
receive automatic updates for plugins that are compatible with their emesene
version. This is what the "required emesene version" field is for. Whether an
update is available will be derived from the "version" field. If you update
a plugin, please increase the "version" number